Add readmes
This provides the overall project readme, which is distinct from the
individual ones included with each respective application. It also
initiates the Quiver application with a readme, as the project's first
theme will be designed for it.

These allow participants to understand the project's purpose, how to
find supporting documentation (e.g., contributing guidelines,
installation instructions, etc.), and access sources that were utilized
to develop the project.