Update contribution guidelines diction
Previously, the tone lacked enough warmth and there were a few typos.
This created a less inviting, as well as less clear message to participants.

Now, the text has a friendly, active-voice tone and the typos are all
corrected. This creates consistency, and an easier to understand message
that encourages participants to keep reading.