Add label convention guidelines
Previously, no guidelines existed outside the default GitHub labels in
the repo's issue section. While the labels could still be used, there
was no direction in how they applied specifically to this project.

Now, there is a dedicated section with the purpose for each type
defined. This allows for consistency, organization, and clarity of issue
types for all project participants.