Add a code of conduct
Without a code of conduct supplied for the project, participants are
left with only assumptions regarding the project's community policies.

This provides a code of conduct, and references to it in other
supporting documentation. This will set clear expectations, and serve as
a behavioral point of reference for all project participants.