Quiver/Calm: Add minified web view highlights
This provides minified versions of the editor, preview, and presentation
web view highlights.

The original non-minified versions are for developing, but the primary
theme file will need a minified json string of each web view highlights.
This allows one file to be read & modified easily, and the other to be
more concise for production needs.